chore: moves edc catalog transformers
