+++ core/control-plane/control-plane-transform/src/main/java/org/eclipse/edc/connector/controlplane/transform/edc/contractagreement/from/JsonObjectFromContractAgreementTransformer.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.edc.connector.api.management.configuration.transform;
+package org.eclipse.edc.connector.controlplane.transform.edc.contractagreement.from;
 
 import jakarta.json.JsonBuilderFactory;
 import jakarta.json.JsonObject;
