+++ core/control-plane/control-plane-transform/src/main/java/org/eclipse/edc/connector/controlplane/transform/edc/contractdefinition/from/JsonObjectFromContractDefinitionTransformer.java
@@ -13,7 +13,7 @@
  *
  */
 
-package org.eclipse.edc.connector.controlplane.api.management.contractdefinition.transform;
+package org.eclipse.edc.connector.controlplane.transform.edc.contractdefinition.from;
 
 import jakarta.json.JsonBuilderFactory;
 import jakarta.json.JsonObject;
