+++ core/control-plane/control-plane-transform/src/test/java/org/eclipse/edc/connector/controlplane/transform/edc/catalog/to/JsonObjectToCatalogRequestTransformerTest.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.edc.connector.controlplane.api.management.catalog.transform;
+package org.eclipse.edc.connector.controlplane.transform.edc.catalog.to;
 
 import com.apicatalog.jsonld.document.JsonDocument;
 import jakarta.json.Json;
