+++ core/control-plane/control-plane-transform/src/test/java/org/eclipse/edc/connector/controlplane/transform/edc/contractagreement/from/JsonObjectFromContractAgreementTransformerTest.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.edc.connector.api.management.configuration.transform;
+package org.eclipse.edc.connector.controlplane.transform.edc.contractagreement.from;
 
 import jakarta.json.Json;
 import jakarta.json.JsonObject;
