+++ core/control-plane/control-plane-transform/src/test/java/org/eclipse/edc/connector/controlplane/transform/edc/contractdefinition/from/JsonObjectFromContractDefinitionTransformerTest.java
@@ -13,7 +13,7 @@
  *
  */
 
-package org.eclipse.edc.connector.controlplane.api.management.contractdefinition.transform;
+package org.eclipse.edc.connector.controlplane.transform.edc.contractdefinition.from;
 
 import jakarta.json.Json;
 import jakarta.json.JsonBuilderFactory;
