+++ core/control-plane/control-plane-transform/src/test/java/org/eclipse/edc/connector/controlplane/transform/edc/contractdefinition/to/JsonObjectToContractDefinitionTransformerTest.java
@@ -13,7 +13,7 @@
  *
  */
 
-package org.eclipse.edc.connector.controlplane.api.management.contractdefinition.transform;
+package org.eclipse.edc.connector.controlplane.transform.edc.contractdefinition.to;
 
 import jakarta.json.JsonObject;
 import jakarta.json.JsonObjectBuilder;
