+++ extensions/control-plane/api/management-api/catalog-api/build.gradle.kts
@@ -25,6 +25,7 @@ dependencies {
     implementation(project(":core:common:lib:validator-lib"))
     implementation(project(":extensions:common:api:lib:management-api-lib"))
     implementation(project(":extensions:common:http:lib:jersey-providers-lib"))
+    implementation(project(":core:control-plane:control-plane-transform"))
 
     implementation(libs.jakarta.rsApi)
