+++ extensions/control-plane/api/management-api/contract-definition-api/build.gradle.kts
@@ -21,11 +21,12 @@ dependencies {
     api(project(":spi:control-plane:control-plane-spi"))
     api(project(":spi:common:transaction-spi"))
     api(project(":spi:common:participant-context-single-spi"))
-    
+
     implementation(project(":extensions:common:api:lib:management-api-lib"))
     implementation(project(":core:common:lib:api-lib"))
     implementation(project(":core:common:lib:validator-lib"))
     implementation(project(":extensions:common:http:lib:jersey-providers-lib"))
+    implementation(project(":core:control-plane:control-plane-transform"))
 
     implementation(libs.jakarta.rsApi)
